requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
Pillow==10.0.0
moviepy==1.0.3
numpy==1.24.3
imageio==2.31.1
imageio-ffmpeg==0.4.8